Removed the reddit_auto cog
