+++ config.yaml
@@ -8,7 +8,6 @@ Bot:
         - garlic
         - pokemon
         - reddit
-        - reddit_auto
         - misc
 
 Anagram:
